Add files via upload
R code, extremely naive and poorly written, to analyze the data. It has comments so it should be possible to use it.
Assumes that the files have the same structure, with first column the main taxon, column #2 Genus_epithet, column #3 an ID of the observer, column #4 the year of the observation, columns 5 and 6 Longitude and Latitude, columns 7-10 the ID of the hexagon at which the observation belongs, and column 11 the identity of the Potential Vegetation, as per Rzedowsky, 1986 Vegetacion de Mexico.